[Feat] 핵심 기능 구현 및 버그 수정
주요 변경사항:
- 인증/회원가입 로직 검증 및 버그 수정 (JWT secret, nickname 중복 체크 등)
- 홈 화면 기능 구현 (다른 사용자 프로젝트 표시, 모집 현황)
- 파트너 매칭 개선 (역할 검증 추가, acceptApplication 제거)
- 결제 로직 수정 (직군별 고정 가격, 시나리오별 자동 승인)
- Gemini Task 자동 생성 기능 추가 (프로젝트 확정 시)
- Task 수정 시 ToDo 업데이트 기능 추가
- 프로필 이미지 조회/수정 기능 추가
- CLAUDE.md 추가 (프로젝트 가이드)

+++ src/main/java/com/example/off/domain/pay/controller/PayController.java
@@ -40,7 +40,7 @@ public class PayController {
     }
 
     private Long getMemberId(HttpServletRequest req) {
-        Object attr = (Long) req.getAttribute("memberId");
+        Object attr = req.getAttribute("memberId");
         if (attr instanceof Long id) return id;
         throw new OffException(ResponseCode.INVALID_TOKEN);
     }
